backend/pyproject.toml
[project]
name = "claude-agent"
version = "0.1.0"
description = "Claude Agent with OpenAI-compatible API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.6",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
